Move queues into its own file
